Consolidate particle MCMC files, drop redundant fork step, add notes
- Inline traced_rng.jl into particle_mcmc.jl as a section (it must precede
  Particle/PGState, which name TracedRNG in their types).
- Merge the container unit tests into test/mcmc/particle_mcmc.jl and remove the
  now-empty essential test group.
- Drop the set_taped_globals! call in fork: deepcopy already preserves the
  task-particle back-reference (verified against the full particle suite).
- Add WHY notes for the might_produce hints and the reference RNG rewind.
